+++ travel-api/reservation-service/src/main/java/org/microarchitecturovisco/reservationservice/domain/dto/HotelInfo.java
@@ -4,11 +4,12 @@ import lombok.Builder;
 import lombok.Data;
 
 import java.util.Map;
+import java.math.BigDecimal;
 
 @Data
 @Builder
 public class HotelInfo {
     private String name;
-    private Float hotelPrice;
+    private BigDecimal hotelPrice;
     private Map<String, Integer> roomTypes;
 }
